refactor: :card_file_box: Modified Balance field data type to decimal

+++ Models/Account.cs
@@ -3,7 +3,7 @@ namespace BankApp.Models
     public class Account
     {
         public int AccountId { get; set; }
-        public float Balance { get; set; }
+        public decimal Balance { get; set; }
         public string Number { get; set; } = null!;
         public int CustomerId { get; set; }
         public Customer Customer { get; set; } = null!;
